refactor(Resources): fix imports
