+++ packages/client/src/game/services/Resources/helpers.ts
