+++ packages/client/src/game/services/index.ts
@@ -1,4 +1,5 @@
 export { Controller } from './Controller';
 export { Game } from './Game';
+export { resources } from './Resources/Resources';
 export { View } from './View';
 export { Zone } from './Zone';
