+++ packages/client/src/game/utils/index.ts
@@ -1,3 +1,2 @@
 export { EventEmitter } from './EventEmitter';
-export { getAssetType } from './getAssetType';
 export { isKeyOfObject } from './isKeyOfObject';
